[MINOR] fix a few remaining printf-like formats on 64-bit platforms
Mainly two sizeof() returning size_t which is not the size of an int
on 64-bit platforms.

+++ src/cfgparse.c
@@ -3091,7 +3091,7 @@ int readcfgfile(const char *file)
 			 * Watch out for the last line without the terminating '\n'!
 			 */
 			Alert("parsing [%s:%d]: line too long, limit: %d.\n",
-				file, linenum, sizeof(thisline)-1);
+			      file, linenum, (int)sizeof(thisline)-1);
 			goto err;
 		}
